requirements.txt
streamlit
PyPDF2
spacy
scikit-learn
python-docx
matplotlib